Trading Bot/trading_bot/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
requests>=2.31.0
schedule>=1.2.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
ta>=0.11.0
MetaTrader5>=5.0.45
pytz>=2023.3
joblib>=1.3.0
scipy>=1.11.0
